requirements.txt
Django>=5.2
djangorestframework>=3.16
djangorestframework-simplejwt>=5.5
django-cors-headers>=4.9
PyJWT>=2.7
